Adds basic infraestructure for video conversion
This commit adds a VideoConversionService which will be an asynchronous
service for converting between video formats upon upload. Videos that
were uploaded, but are being converted, are returned from the backend with
a flag attribute in JSON ("is-processing"=true)

This commit does not implements the Service, only a dummy implementation
is provided.