Changes the primary key of VideoModification
Changes VideoModification to be able to repeat a modifier in a video. The
new key is an integer and it is verified that the video, modifier, start
and end data are not already registered.
When removing VideoModificationId, is necessary to redo the delete
functionality. Now removed by id and is necessary to add the functionality
to obtain a VideoModification by id to delete the object.