Creates a process to clean the videos on the disc
Configures a task that is executed daily to verify that the stored videos are those
that are in the database. If there is a video that is not in the database, these
video are removed from the disk. For this, it is necessary to change the
configuration of the Wildfly server.