Fixes DB integrity check errors upon deleting explorations or patients
When an exploration or a patient is deleted, its associated videos
are removed via JPA cascade removals, configured in @OneToMany
associations. Since there was no @OneToMany relation between
Video and VideoModification, a manual removal of VideoModification
was implemented in DefaultVideoDAO. This works as expected when
a video is directly removed by using DefaultVideoDAO. However,
if the video is deleted indirectly through a cascade removal operation
coming from patient or exploration, this manual removal is not
executed, leading to integrity check errors then the removing video
has VideoModifications.

This commit creates a private @OneToMany relation between Video
and VideoModification in order to both fix the bug and keep this bidirectional
inaccessible from higher layers. The manual removal of VideoModification
in DefaultVideoDAO is no longer required and thus it is removed.