Adds Image entity, the resource and service to upload it
This commit creates an Image entity, which is related to a Video and a Gallery.
The Video is kept unrelated to Image, whereas Image and Gallery are bidirectionally
related.

Updates the Gallery service to obtain a gallery given an identifier.

Adds a new resource and service to upload images.

Finally, the file storage and the periodic storage cleaner are adapted to
take into account image files.