Adds pagination to explorations
The function responsible for returning the patient's explorations is
eliminated. Now, all explorations are obtained from the same function in
which it is verified if a patient is passed and then the corresponding
query is sent to the database. This function also receives the page size
and the current page. With this data, the index of the first and the
last exploration that will be shown are calculated, with them a
ListingOptions object will be created that will help to return the
necessary explorations.

Two new functions are added that return the number of total explorations
and the number of explorations of a given patient.

In addition, the “X-Pagination-Total-Items” header is allowed so that it
can be read in the frontend.