Adds pagination to the images
Now the images of a gallery are searched by the gallery's identifier
in order to paginate the images. Therefore, the bidirectional relationship
between Gallery and Image is removed.

Also adds the fps attribute to the Video entity to know the frame per
second of the video and be able to calculate the frame number of the
image.