Checks that a polyp recording is created in an available time range
Now, multiple polyp recordings can be added to the same polyp within the
same video. To do this, ID field is added to the polyp recording, becoming
the primary key. Because of this, the class polypRecordingId is removed
since it is unnecessary.

The delete method has been modified. Now, the argument sent is a polyp
recording instead of the video and the polyp. For this it was necessary
to create the method to obtain a polyp recording by means of the id.