Adds the deletion of modifiers
Adds the functionality to remove an existing modifier. Only the ADMIN
role can do this deletion. If the modifier has already been assigned to
a video, an exception is thrown that can not be removed. To throw this
exception, it was necessary to create a function that returns all video
modifications given a modifier.