Adds the fields number of polyps and number of videos to ExplorationData
To show the number of videos and polyps, it is necessary to add the
attributes "number of polyps" and "number of videos" to ExplorationData
(what the frontend will receive). Only getters are created for these
attributes. To obtain this data, two methods are created in the
Exploration entity that indicates the sizes of the polyps and videos
lists.

In addition, the name of the explorations attribute "date" has been
changed by "explorationDate".