Adds a filter to select the images in the gallery
This commit adds a filter to indicate if it want to see the images
located, not located or all.

A resource is also added to obtain the image identifiers according
to the filter. This improves the loading of images when the polyps
are located in the images.

The URL that returns the images is changed to follow the REST
convention. Instead of being '/image/gallery/{gallery_id}' it
becomes in 'image?gallery={gallery_id}".

The queries to the database are optimized to obtain only the
necessary data. These queries should be changed to Criteria.