Adds a new filter to get the not located images with polyp
This commit adds a new filter "not_located_with_polyp" to obtain the
images that have a polyp and are not located. In addition, a filter
(called "with_polyp") is also added to obtain the total number of images
with an assigned polyp. This number is sent through the header
"X-With-Polyp-Total-Items".