Adds the creation of zip files to download the images of the galleries
This commit adds the functionality of preparing the zip with the desired
gallery images that will be downloaded. For that, it uses
StreamingOutput (which allows data to be sent fragmentedly) and
ZipOutputStream. When creating the zip, the desired filter is checked
and if the location of the polyp should be drawn on the image or not. In
case of drawing it, the exception that occurs when accessing the
location of an image that does not have it is controlled, to continue
executing the code. The name used for the gallery is the id, while the
name of the images is composed of the polyp id (if applicable), the
video id and the frame to which it belongs.

To do this, a new public resource called DownloadResource is created
since the download request is requested without authentication
credentials.