Adds password recovery and email sending for it
This commit adds the password recovery system. For it, there are some
new classes and entities in all the layers of the project.
There are two possible flows. First, backend receives a petition from
the user to create a recovery. User can solicite it using his login or
email. The system creates it with actual date, store it and then send an
email to the user with a link, that user may use to create the new
password.
After that, in second flow, user  create a new password. When the
backend receives it, it checks that there is a recovery request for this
user, and that the date is not more than one day ago. If that's correct,
stores the new password, removes the request and sends a confirmation
message. If date is away but request still in database, removes the
request and sends an error message.
Also, there is an scheduled service every day at midnight that takes
care of removing recovery requests from database when their dates go
over.