Adds backend functions to confirm different entities
Adds functions to confirm Exploration, Polyp, all Polyps from a
exploration, VideoModification and PolypRecording.

The changes done to VideoModification also include support for
editing this entity.

The confirmation of all polyps and video modifications of an
exploration when an exploration is confirmed. All polyp recordings
that belong to a polyp are also confirmed by confirming the polyp.

In addition, it is verified that the different elements can be confirmed
and if it does not return a forbidden error.