Fixes bug preventing from deleting a polyp with images
A bug was causing an error when a polyp with an image in a gallery was
deleted. Due to the image not being deleted, an integrity error was
produced in the database.

This commit fixes this bug an now polyps should be deleted even if the
have an associated image.