Optimizes and refactorizes images in galleries queries
The queries to count the number of images in the galleries were
recovering the full list of images and then counting them in Java
code. This is very slow, as the entities for all the images must be
created, although they are not effectively used.

This commit optimizes this queries by doing the image count in the
database (i.e. using "count" SQL queries).

In addition, the "filter" parameter used in some related methods is
now an enum instead of a plain String.