Fix "Cannot access default fields of Properties" error in latest Eclipse
The following error happens when trying to open the REST project in the
latest Eclipse versions (2022-03):

https://stackoverflow.com/questions/67221463/cannot-access-defaults-field-of-properties

This error may be caused due to Eclipse moving to a newer Java version
internally, while this project still depends on Java 8.

This problem was fixed for me by following the StackOverflow answer of
fixing the version of the Maven WAR plugin to something modern, which is
a thing that makes sense to do anyway.