Add original video storage feature
This feature, which can be configured via the WildFly
java:global/piba/defaultfilestorage/original option, allows PIBA to keep
the videos submitted by users as-is in a certain directory. If that
option is set to "DISABLED" this new feature will be disabled.

The user-submitted videos are stored with the original filename
indicated by their user agent, when possible. To deal with rogue clients
and filesystem differences, file names are sanitized against invalid
characters and path traversal sequences, to avoid allowing clients to
write files outside of the intended place.

The storage cleaner was updated to also clean up original orphan files,
and made more precise in the files it deletes by considering normal
files with a UUID name only.

The changes were tested by uploading files that contained emojis in
their file names under several WildFly option configurations.